requirements.txt
azure-core==1.30.1
azure-identity==1.15.0
chromadb==0.4.24
huggingface-hub==0.21.3
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.39
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
nltk==3.8.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
streamlit==1.33.0
openai==1.16.1
unstructured==0.12.5
unstructured-client==0.21.0
unstructured-inference==0.7.23